+++ src/resources/containers/instances/index.ts
@@ -1,7 +1,6 @@
-import * as Volumes from "./volume";
-
+export * as Telemetry from "./telemetry";
+export * as Volumes from "./volume";
 export * from "./instance";
 export * from "./console";
 export * from "./ssh";
 export * from "./tasks";
-export { Volumes };